rename to `ControlledViaAnd`
