--- qualtran/bloqs/mcmt/multi_controlled_bloq.py
@@ -1,116 +1,0 @@
-#  Copyright 2024 Google LLC
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      https://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-from functools import cached_property
-from typing import TYPE_CHECKING
-
-import numpy as np
-from attrs import frozen
-
-from qualtran import Bloq, Controlled, CtrlSpec
-from qualtran.bloqs.basic_gates import XGate
-from qualtran.bloqs.mcmt.ctrl_spec_and import CtrlSpecAnd
-
-if TYPE_CHECKING:
-    from qualtran import BloqBuilder, Signature, SoquetT
-    from qualtran.resource_counting import BloqCountT, SympySymbolAllocator
-
-
-@frozen
-class MultiControlledBloq(Bloq):
-    ctrl_spec: CtrlSpec
-    subbloq: Bloq
-
-    @cached_property
-    def signature(self) -> 'Signature':
-        return Controlled(self.subbloq, self.ctrl_spec).signature
-
-    @cached_property
-    def ctrl_reg_names(self) -> tuple[str, ...]:
-        return tuple(Controlled(self.subbloq, self.ctrl_spec).ctrl_reg_names)
-
-    def _is_single_bit_control(self) -> bool:
-        return self.ctrl_spec.num_qubits == 1
-
-    @cached_property
-    def _single_control_value(self) -> int:
-        assert self._is_single_bit_control()
-        return self.ctrl_spec._cvs_tuple[0]
-
-    def build_composite_bloq(self, bb: 'BloqBuilder', **soqs: 'SoquetT') -> dict[str, 'SoquetT']:
-        # compute the control bit
-        if self._is_single_bit_control():
-            ctrl_soqs = None
-            q = soqs.pop(self.ctrl_reg_names[0])
-            if self.ctrl_spec.shapes[0] == (1,):
-                assert isinstance(q, np.ndarray)
-                q = q[0]
-
-            if self._single_control_value == 0:
-                q = bb.add(XGate(), q=q)
-        else:
-            # map input control soqs to the CtrlSpecAnd controls
-            ctrl_spec_and_bloq = CtrlSpecAnd(self.ctrl_spec)
-            ctrl_soqs = {
-                inner_ctrl_reg.name: soqs.pop(name)
-                for name, inner_ctrl_reg in zip(
-                    self.ctrl_reg_names, ctrl_spec_and_bloq.control_registers
-                )
-            }
-
-            # compute the control bit
-            ctrl_soqs = bb.add_d(CtrlSpecAnd(self.ctrl_spec), **ctrl_soqs)
-            q = ctrl_soqs.pop('target')
-
-        # add single-controlled-subbloq
-        _, adder = self.subbloq.get_ctrl_system(CtrlSpec())
-        out_q_iter, out_soqs_iter = adder(bb, [q], soqs)
-        (q,) = tuple(out_q_iter)
-        soqs = dict(zip([reg.name for reg in self.subbloq.signature], out_soqs_iter))
-
-        # uncompute the control bit
-        if self._is_single_bit_control():
-            if self._single_control_value == 0:
-                q = bb.add(XGate(), q=q)
-
-            if self.ctrl_spec.shapes[0] == (1,):
-                q = np.array([q])
-            soqs[self.ctrl_reg_names[0]] = q
-        else:
-            # uncompute the control bit
-            ctrl_spec_and_bloq = CtrlSpecAnd(self.ctrl_spec)
-            assert ctrl_soqs is not None
-            ctrl_soqs['target'] = q
-            ctrl_soqs = bb.add_d(ctrl_spec_and_bloq.adjoint(), **ctrl_soqs)
-
-            # map the control soqs back to the original bloq's control soqs
-            soqs |= {
-                name: ctrl_soqs[inner_ctrl_reg.name]
-                for name, inner_ctrl_reg in zip(
-                    self.ctrl_reg_names, ctrl_spec_and_bloq.control_registers
-                )
-            }
-
-        return soqs
-
-    def build_call_graph(self, ssa: 'SympySymbolAllocator') -> set['BloqCountT']:
-        counts: set['BloqCountT'] = {(self.subbloq.controlled(), 1)}
-
-        if self._is_single_bit_control():
-            if self._single_control_value == 0:
-                counts |= {(XGate(), 2)}
-        else:
-            ctrl = CtrlSpecAnd(self.ctrl_spec)
-            counts |= {(ctrl, 1), (ctrl.adjoint(), 1)}
-
-        return counts
